% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_function.R
\name{anova.linear_regression}
\alias{anova.linear_regression}
\title{ANOVA for Linear Regression}
\usage{
\method{anova}{linear_regression}(object)
}
\arguments{
\item{object}{An object of class \code{linear_regression} created by the \code{linear_regression} function.}
}
\value{
A summary table of the ANOVA analysis.
}
\description{
This function performs an ANOVA (Analysis of Variance) for a linear regression model.
}
\examples{
model <- linear_regression(X, y)
anova(model)

}
