% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linear_regression}
\alias{summary.linear_regression}
\title{Summary for Linear Regression model}
\usage{
\method{summary}{linear_regression}(object)
}
\arguments{
\item{object}{An object of class \code{linear_regression} created by the \code{linear_regression} function.}
}
\value{
A list containing:
\item{Call}{The matched function call.}
\item{Residuals}{Summary statistics of the residuals.}
\item{Coefficients}{Estimated regression coefficients.}
\item{t.value}{T statistic value}
\item{pr...t..}{P value}
\item{Residual standard error}{An estimate of the residual standard deviation.}
\item{R-squared and Adjusted R-squared}{Measures of the goodness of fit of the model.}
}
\description{
This function computes and displays diagnostic statistics for a linear regression model
fitted using the \code{linear_regression} function. It provides information about the model's
residuals, coefficients, and goodness of fit.
}
\examples{
# Generate example data
set.seed(47)
X <- cbind(rnorm(100), rnorm(100))
y <- 2 * X[, 1] - X[, 2] + rnorm(100)
model <- linear_regression(X, y)
summary(model)
}
